refactor(grid): add dragger plugin
